refactor(functions): separate twUsers utils
